fix: changed datatzpe of g_oneMegabyte to size_t and use this var for comparison

+++ inc/constants.hpp
@@ -4,9 +4,10 @@
 
 /* ====== DECLARATIONS ====== */
 
+#include <cstddef>
 namespace constants {
 
 extern const int g_decimalBase;
-extern const int g_oneMegabyte;
+extern const size_t g_oneMegabyte;
 
 } // constants
